Solution to Check Triangle Type

+++ Check Triangle Type@@6788dd63490637cfc44e1b97/code.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 int main() {
-    float side1,side2,side3;
+    int side1,side2,side3;
     scanf("%d %d %d",&side1,&side2,&side3);
     if (side1 == side2 && side2 == side3 && side1 == side3){
         printf("Equilateral");
